CodeRunner maintains a `queue` promise
The previous commit produced a Promise waiting for all previous jobs to
resolve. That made me a bit uneasy, because it feels like it's doing too
much: you only need to wait for the last task on the queue to finish,
and by induction all the earlier ones have too.

This changes the queuing method to store a single promise in the `queue`
property, and replace it with a new promise each time a task is enqueued.

It looks like this method is significantly faster than the other method,
but more than I expected. I assumed the JS engine would optimise away
the big list of promises, but maybe it doesn't.