show errors in the postamble
fixes #14

If there's an error in the postamble, then a message is shown to the
student. If "show stderr" is turned on, then the error traceback is
shown.

This also adds marking notes `postamble_stderr` and `preamble_stderr`,
so you can see the error message in the editor's marking testing tab.