setAnswerJSON only uses the most recent value
fixes #19

When replacing selected text, there are two 'change' events: one removes
the selected text, and another adds in the new text.

This means that `setAnswerJSON` is called twice. Because `setAnswerJSON`
is async, the part that compares the editor's content with the given
value happens after both changes have been applied.

Only the final call to `setAnswerJSON` actually needs to do anything, so
this commit adds a check that the `answerJSON` value after the `await`
is the same as the `answerJSON` given to the function. If not, the
function returns immediately. For the final value, the two variables
refer to the same object so the function continues.