use WebR 0.2.1
It works a lot better, and is hosted by r-wasm.org instead of George's
jsdelivr repo.

WebR now returns real values, and produces a list of lines of output.

It's assumed that images will be saved as SVG plots. WebR has a canvas
output, but I'd prefer to get an SVG.

To produce an SVG, first set up svglite:

    webr::install("svglite")
    svg()

and then after plotting close the device to write an .svg file:

    dev.off()