webR: only include an image in result if it's non-empty
When you open the `svg()` device, it creates a file `Rplot001.svg`, but
hasn't written anything to it yet. If the plot is produced in a later
code block, then the code that looks for images will see the file before
it has anything in it.

There's no way that I can see of checking if a file is open, so this
just waits for an image file's contents to be non-empty.