use contexts for evaluating code
To solve the problem of different questions preloading different sets of
files, this commit introduces a system of contexts.
When preloading a runner or running code, you can give a context ID.
This is a namespace for the files in the `/resources` directory.
There's a default context, for backwards compatibility - the list of
files available in a given context is the union of the default context's
files and the named context's files.