use `captureGraphics: false` when running R code
webR v0.3.0 introduced a new default behaviour of capturing plots. The
release notes say this is only supposed to capture plots to the canvas
device, but that might just be awkward wording.

In order to save SVG plots to webR's filesystem, we need to turn this
behaviour off.

This commit also updates webR to 0.4.2