replace Ace with CodeMirror 6
The old ace editor didn't work nicely with the new `<numbas-exam>`
element - I think it probably assumes that the element lives in the main
DOM.

This commit replaces ace with a CodeMirror editor 6 component,
`<code-editor>`. I'm not sure I want to put the source for that here,
without rearranging the repo so the finished extension files are in a `dist` folder.